fix: Add missing torch import in overlap detector
- Fixed 'name torch is not defined' error in overlap_detector.py
- Added torch import to optional imports section
- Added check for PYANNOTE_AVAILABLE before using torch
- This fixes the error that occurs during audio enhancement with secondary speaker detection

+++ tests/test_huggingface_schema_complete.py
@@ -103,7 +103,7 @@ class TestHuggingFaceSchemaComplete(unittest.TestCase):
             "transcript": Value("string"),
             "length": Value("float32"),
             "dataset_name": Value("string"),
-            "confidence_score": Value("float64")
+            "confidence_score": Value("float32")
         })
         
         # Create dataset with sample data
@@ -127,7 +127,7 @@ class TestHuggingFaceSchemaComplete(unittest.TestCase):
                 "transcript": Value("string"),
                 "length": Value("float32"),
                 "dataset_name": Value("string"),
-                "confidence_score": Value("float64")
+                "confidence_score": Value("float32")
             })
             
             dataset = Dataset.from_list([self.sample_data], features=features)
